Add files via upload
Federal Elections Commission finance data. If I had been smarter, I would have started here to get my candidates list. It includes a status marker for the general election and a unique identifier for each candidate. There are two caveats with the data--some people register for multiple races under slightly different names, and there is a lag to the reporting, so some data are incomplete at the time of the election.